* gpx-tracks:
  Workaround to improve the reliability of a test
  Fixed recently-introduced compilation error in local unit tests
  Load geo: URLs for Google Maps, Gaia GPS, Google Earth, and others...
  Zoom in on track when loaded
  Show all details about a selected GPX track
  Fully implemented working track fragment to show a map with the track drawn on it
  Upgraded to the latest Kotlin 1.7.x
  Bump most dependencies to the lastest version available
  Updated to-do list
  Added an alert dialog to select a GPX track to view

+++ app/src/androidTest/java/org/northwinds/app/sotachaser/ui/regions/RegionFragmentTest.kt
@@ -202,7 +202,7 @@ class RegionFragmentTest {
             if (view == null)
                 throw noMatchingViewException
             val recyclerView = view as RecyclerView
-            assertThat(recyclerView.adapter?.itemCount,
+            assertThat(recyclerView.adapter?.itemCount as Int,
                 Matchers.`is`(Matchers.greaterThanOrEqualTo(10))
             )
         }
